literature_review/requirements.txt
pyautogen>=0.2.0
python-dotenv
arxiv
perplexityai
requests
